Fix #663, more consistent use of data types

+++ fsw/cfe-core/src/inc/cfe_tbl.h
@@ -116,7 +116,7 @@ typedef enum CFE_TBL_SrcEnum
 /** \brief Table Info */
 typedef struct CFE_TBL_Info
 {
-    uint32                Size;                             /**< \brief Size, in bytes, of Table */
+    size_t                Size;                             /**< \brief Size, in bytes, of Table */
     uint32                NumUsers;                         /**< \brief Number of Apps with access to the table */
     uint32                FileCreateTimeSecs;               /**< \brief File creation time from last file loaded into table */
     uint32                FileCreateTimeSubSecs;            /**< \brief File creation time from last file loaded into table */
@@ -265,7 +265,7 @@ typedef struct CFE_TBL_Info
 **/
 CFE_Status_t CFE_TBL_Register(CFE_TBL_Handle_t *TblHandlePtr,                  /* Returned Handle */
                           const char   *Name,                              /* Application specific name  */
-                          uint32  Size,                                    /* Size, in bytes, of table   */
+                          size_t  Size,                                    /* Size, in bytes, of table   */
                           uint16  TblOptionFlags,                          /* Tbl Options Settings     */
                           CFE_TBL_CallbackFuncPtr_t TblValidationFuncPtr); /* Ptr to func that validates tbl */
